test: select operator e2e test
